windows: clean quirks
https://github.com/warmcat/libwebsockets/issues/3039

+++ lib/roles/http/server/server.c
@@ -212,7 +212,7 @@ done_list:
 	for (m = 0; m < limit; m++) {
 
 		if (a->info && a->info->vh_listen_sockfd)
-			sockfd = dup(a->info->vh_listen_sockfd);
+			sockfd = dup((int)a->info->vh_listen_sockfd);
 		else
 			sockfd = lws_fi(&a->vhost->fic, "listenskt") ?
 					LWS_SOCK_INVALID :
